requirements_robot.txt
numpy=1.18.2
serial==0.0.97
smbus2==0.3.0   
RPi.GPIO==0.7.0

